mainline: linux: config, connect network on startup

+++ 06_generate_debian_console_root.sh
@@ -65,7 +65,19 @@ rm -rf ./ignore/.root/etc/systemd/system/multi-user.target.wants/wpa_supplicant@
 #ln -L -f -s -v /lib/systemd/system/resize_filesystem.service --target-directory=./etc/systemd/system/multi-user.target.wants/
 #cd ../../
 
-cp -v ./ignore/.root/etc/bbb.io/templates/eth0-DHCP.network ./ignore/.root/etc/systemd/network/eth0.network || true
+if [ -f ./ignore/.root/etc/bbb.io/templates/eth0-DHCP.network ] ; then
+	cp -v ./ignore/.root/etc/bbb.io/templates/eth0-DHCP.network ./ignore/.root/etc/systemd/network/eth0.network || true
+else
+	echo '[Match]' > ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'Name=eth0' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'Type=ether' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '[Link]' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'RequiredForOnline=yes' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '[Network]' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'DHCP=ipv4' >> ./ignore/.root/etc/systemd/network/eth0.network
+fi
 
 # setuid root ping+ping6
 chmod u+s ./ignore/.root/usr/bin/ping ./ignore/.root/usr/bin/ping6
@@ -82,7 +94,7 @@ echo 'File Size'
 du -sh ignore/.root/ || true
 echo '---------------------'
 
-dd if=/dev/zero of=./deploy/input/root.ext4 bs=1 count=0 seek=2048M
+dd if=/dev/zero of=./deploy/input/root.ext4 bs=1 count=0 seek=2200M
 mkfs.ext4 -F ./deploy/input/root.ext4 -d ./ignore/.root
 
 if [ -f ./.06_generate_root.sh ] ; then
